requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3
schedule==1.2.0
asyncio==3.4.3
flask==2.3.3
werkzeug==2.3.7
aiofiles==23.2.1
pillow==10.0.1
emoji==2.8.0 